Added Home Screen and Product Screen, extracted new Product, Rating components.
Installed react-router-dom for front-end routing, and used it to add routes to
App.js, the '/' route handled by the new component HomeScreen, and the
'/product/:id' route handled by the new ProductScreen component.

Extracted a new Product component to display product information in cards on the
Home screen. Extracted a Rating component to display the rating stars and number
of reviews in the Product component. Added styles to improve the look of the
product details screen.

In the data.js file, added some new fields to the products documents, a 'name'
field which is shown at the top of a product listing. The name field is so that
we can display more than just the title of a book, showing 'Title (Paperback)',
for example.