Made modifications to fetch data from the backend.
Installed the axios package for making AJAX (xmlhttprequest) requests.

Added 'proxy' setting to package.json to redirect AJAX requests to the backend
running on localhost.

Created a LoadingBox component to display a spinner and Loading... message while
the page is loading. Created a MessageBox component to display nicely info,
warning and error messages as needed. Added CSS classes for alert, alert-info
and alert-warning.

Modified the Home screen to use the LoadingBox and MessageBox components, and
to use React hooks for state: loading and error variables, and a products array,
and an Effect hook to run code when our component didMount or didUpdate, to
fetch data for the products to display.