Created new Cart screen and new route that the CartScreen component handles.
The cart screen is in CartScreen.js in the frontend/src/screens directory. For
right now, it is a skeleton screen displaying the information with no styling.

In App.js we added a new <Route /> with a path of '/cart/:id?' and set it to be
handled by the CartScreen component.