Added Add to Cart handler for Product screen, select for choosing quantity.
Added [qty, setQty] as a redux useState variable to the Product screen.

Added addToCartHandler function which uses props.history.push() to redirect the
user to the cart screen.

In the ProductScreen.js, moved the Add to Cart button inside a conditional
rendering construct, so that it only appears when product.countInStock > 0. Also
added a Select element for choosing the quantity to add to the cart.

Also made some styling changes in App.css, to make the Qty select
element’s styling more like that of the buttons, except that it does not change
the mouse cursor to a pointer (that should only be done for buttons and links).