Create constant, action, reducer for adding an item to the shopping cart.
The constant is CART_ADD_ITEM, defined in
frontend/src/constants/cartConstants.js.

The action is dispatched by the addToCart function defined in cartActions.js. We
also use local storage setItem() to set the cartItems object to hold the items
in the user’s cart so that they are not lost if the page is refreshed.

The reducer is the cartReducer defined in cartReducers.js.