Add the cart reducer to the redux store, render badge with count of cart items.
Also the redux store is changed so that its initial state contains a cart object
which has a cartItems element, which is loaded from local storage if it has been
stored there.

Modified App.js to show the item count badge next to the cart in the header of
the App. This badge is styled with the badge class defined in App.css.