Implemented the nicer Shopping Cart screen.
The new CartScreen component lays out the screen in two columns, the left column
is the list of items in the cart, with each item taking up a row. An item row
has a small product image, the product name, a select showing and changing the
item quantity, and a Delete button to remove that item from the cart. The right
column is the action area with the subtotal for price and the Proceed to
Checkout button.

Now we use useSelector to get the cart items from the redux store. We use
conditional rendering to display a MessageBox component saying the cart is empty
if the cartItems array has length 0.

Added a stub function removeFromCartHandler to handle removing an item from the
cart. We will implement this later.

Added checkoutHandler function to handle clicking on the Proceed to Checkout
button. This just redirects the user to the Sign In screen, shipping section.