Add some styling to the item rows in the shopping cart.
Added a CSS ruleset for img.small class which keeps the max-width of the image
thumbnail at 5rem.

Added a CSS min-30 class to apply a minimum width of 30rem to elements. This is
applied to the product name in the item row in the shopping cart screen, to
ensure that as much of the product name is shown as possible.