Create constant, action & reducer for removing an item from the shopping cart.
The new constant is CART_REMOVE_ITEM.

Defined a new function removeFromCart() to dispatch the CART_REMOVE_ITEM action.
This also updates the cart items in local storage.

Defined a new case in cartReducer for CART_REMOVE_ITEM, which returns a new
state object in which the payload product id has been filtered out of the cart
items.