Renamed .eslintrc.js to .eslintrc.cjs to fix LSP error.
LSP complained of an error trying to require() .eslintrc.js when the package it
is in is an ES module and thus should be imported with ’import’. I am not sure
what exactly is going on here, but eslint does use require() statements to load
modules. In any case, renaming .eslintrc.js to .eslintrc.cjs (for Common JS
format) did make the error go away.