Create User model and user router, connect to MongoDB via mongoose. Seed users.
The user router handles all routes that match ’/api/users’. At this point just
added one such route for ’/api/users/seed’, which inserts the users in data.js
file to the User collection in the innerspace database in MongoDB.

In server.js, we use mongoose to connect to the local MongoDB server and
database named innerspace.

app.use() function is registered to handle errors in async express functions and
send the error messages to the frontend for display in the browser. If we don’t
do this, when an error happens such as a duplicate email error when going to the
/api/users/seed URL after having already seeded the database, nothing is shown
in the browser window and it just hangs with the loading spinner spinning in the
tab.