Move product-related routes from server.js to productRouter.js in backend.
In backend/server.js, we remove the route handlers for '/api/products' and
'/api/products/:id', and use app.use() to refer all routes matching
'/api/products' to productRouter, defined in routers/productRouter.js.

The route /api/products is now handled as “/” in productRouter.js and
the /api/products/:id is now the handled as “/:id” in productRouter.js. We also
defined a new handler for the route “/seed”, which is used to seed the MongoDB
database with sample products from the array in data.js.