Created mongoose Product model in backend/models/productModel.js.
The schema for the Product model has different properties for fields related to
books and music products. For example, in the MongoDB a book product will have
the author field filled but not the artist and composer fields. Conversely, a
music product will have the artist and composer fields filled but not the author
field. We will see how well this works as development progresses.

Added a productType field to differentiate between 'Books' products and 'Music'
products.

NOTE: I changed the countInStock field to numInStock. This will necessitate
changes in other source code files.

The publicationDate field is meant to be used for the release date of a music
item.

There is a category field for the category of the product within either the Books
or the Music sections. This notion of category may need to be fleshed out further.
For instance, a book may need to have a category of 'science' but then a sub-
category of 'astronomy'. Similarly for a music item, we may want a category of
'chamber music' and then a sub-category of 'piano'. We shall see.