Changed Product.remove() to Product.deleteMany() in productRouter.
The MongoDB api changed and remove() function is now deprecated; the docs say to
use delete() or deleteMany() instead.

Also added a comment to document that the handler for route '/:id' must be
after the handler for '/seed', as the '/:id' route would match '/seed'
(erroneously) if placed before the /seed handler definition.