Added /api/users/signin route handler to backend in userRouter.
Accepts a user's email and password, and if a user exists in the database
with the given email whose password matches the given password, authenticates
that user be returning the user's info along with a JSON Web Token.

Implemented generateToken() function in backend/utils.js, which given a user,
returns a JSON Web Token for the user.