In backend server.js, fixed bug where error middleware wasn’t being called.
Added a fourth parameter ’next’ to the error handling middleware function passed
to app.use. It was being registered like so:

  app.use((err, req, res) => { … }

and we changed it to:

  app.use((err, req, res, next) => { … }

If a middleware function does not have four parameters, Express or Node will not
recognize it as an error-handling middleware. This is why our handler wasn’t
being called.