Added CheckoutSteps component along with Shipping Address screen.
Created new file CheckoutSteps.js with CheckoutSteps React component, which
displays the steps along the top edge of the screen and indicates which have
been completed and where the user is in the process. Also added CSS styles for
them in App.css.

Created a new ShippingAddressScreen React component in
screens/ShippingAddressScreen.js.

To App.js, added a <Route /> for the new ShippingAddressScreen component.

In cartActions.js, added a new action to save the shipping address. Created a
new constant in cartConstants.js, CART_SAVE_SHIPPING_ADDRESS for this action.

In cartReducers.js, modified the cartReducer function by adding a case for
saving the shipping address.

In the Redux store, in store.js, added shippingAddress to the store, and code
to load the shipping address from local storage.