Implemented placing an order, w/ Place Order screen and Order Details screen.
Created a Mongoose schema and model for an Order in the database.

In backend, created orderRouter for order-related routes. In
orderRouter, implemented handlers for two new routes: '/api/orders'
and '/api/orders/:id'. For '/api/orders' the new handler responds to
POST requests to '/', calls isAuth middleware, and then uses
expressAsyncHandler to save the new order in the database and respond
to the frontend with an appropriate message.

For '/api/orders/:id' the handler responds to GET requests to '/:id',
calling isAuth middleware, and then calling expressAsyncHandler to
look up the order in the database and send the order to the frontend
if found or an error message if the order was not found.

We implemented the isAuth() middleware function in backend/utils.js.
This function checks the request headers for an Authorization header
and if there is one, verifies that it contains a Bearer token that is
valid. If the token is invalid, it sends a 401 error response to the
frontend, otherwise if the token is a valid token for a user, it sets
the 'user' field in the request to the authenticated user and calls
the next middleware function.

In server.js, mounted orderRouter at '/api/orders'.

Added <Route> s to App.js for '/placeorder', handled by component
PlaceOrderScreen, and '/order/:id', handled by component OrderScreen.

Added constants, actions and reducers for ORDER_CREATE_* and
ORDER_DETAILS_*.

Also added cart constant and reducer for CART_EMPTY, to empty the
shopping cart when an order is placed.

Implemented the createOrder() action and detailsOrder() actions.

Created PlaceOrderScreen and OrderScreen.

Added orderCreateReducer and orderDetailsReducer to the Redux store,
store.js.