Added PayPal Buttons to the Order Screen.
Installed npm package ’react-paypal-button-v2’ in frontend project.

The PayPalButton component in this package uses the PayPal Javascript SDK.
We load the PayPal JS SDK via a <script> tag in the Order Screen. This <script>
tag is created in JavaScript code in the useEffect hook.

We need to pass the Client ID of our PayPal App in the URL of the ’src’
attribute of the <script> tag, so that our PayPal buttons settle payments to the
proper PayPal account. We are using PayPal sandbox accounts for the merchant,
InnerSpace, and the buyers.

During development, we are passing a Sandbox Client ID. This will be changed to
a Live Client ID when going into production.

We saved the PayPal Client ID in the .env file in the project root. Since we
need access to this from the frontend code, we use an API call to the backend to
get it.

Set up backend API to retrieve the PayPal Client ID. In the server.js file, add
a handler for GET requests to the ’/api/config/paypal’ endpoint. This handler
sends a response containing the value of the PAYPAL_CLIENT_ID environment
variable, or ’sb’ (for sandbox) if the variable is not defined.

Fixed a bug in the order details reducer which caused the Order Screen to get
stuck in perpetual Loading… from the default value of the ’state’ parameter to
the orderDetailsReducer() function, remove the order: {} key/value pair.

On the Order Screen, added JSX to conditionally render the <PayPalButton>
component if there is an order loaded that is not paid and the sdk is ready.

Added a stub function, successPaymentHandler(). This will handle re-rendering
the screen upon successful payment.