Implemented handling the result returned by PayPal for payment processing.
Added paymentResult field to the Order model in Mongoose:
  paymentResult: {
    id: String,
    status: String,
    update_time: String,
    email_address: String,
  }

In backend orderRouter.js, created a new handler for PUT requests to the
’/api/orders/:id/pay’ endpoint. This looks up the order with ID :id in the
database and sets its isPaid field to true, paidAt to the current date and time,
and its (newly added) paymentResult field to the information returned by PayPal
with a successful payment response.

Created a new payOrder action in orderActions.js.

Created a new order pay reducer in orderReducers.js.

Modified the Order Screen successPaymentHandler() function to accept
paymentResult parameter (from PayPal) and dispatch the payOrder action. Changed
the useEffect() handler to dispatch detailsOrder not only when order is false,
but also when successPay is true or order is true BUT the order id doesn’t match
the orderId of the URL. Also added a new dispatch of ORDER_PAY_RESET just before
dispatching detailsOrder, to prevent an infinite loop causing the page to
refresh again and again.

Surrounded the <PayPalButton> in an empty tag with siblings to show a loading
box if loadingPay is true and the error if errorPay is true.

Added a new CSS style for class alert-success, which is used on the Order Screen
to format the message indicating that the order has been paid.