Added ability for user to see their Order History.
In backend, implemented a handler for GET requests to
'/api/orders/mine' endpoint. This looks up orders in the database that
have a user id matching the current user.

In App.js, added a frontend Route for '/orderhistory' endpoint handled
by the newly created OrderHistoryScreen component.

Implemented the OrderHistoryScreen component. This dispatches the
listOrderMine action in the useEffect hook, and uses the orderMineList
reducer to get the list of orders.

Implemented a new listOrderMine action, along with concommitant
constants and the orderMineList reducer. Added the orderMineList
reducer to the combined reducer in the Redux store.js.