Added ability for user to view their User Profile.
In backend, implemented a handler for GET requests to the
'/api/users/:id' endpoint. This looks up the current user by ID in the
database and sends the user info to the frontend.

In App.js, implemented a frontend Route for the endpoint '/profile'
handled by the ProfileScreen component. Also added a Link to the User
Profile in the dropdown in the header of the app.

Implemented the new ProfileScreen component. This dispatches the
detailsUser action in the useEffect hook, and uses the userDetails
reducer to get the user's info.

Implemented a new detailsUser action, along with concommitant
constants and userDetailsReducer reducer. Added userDetails reducer to
the combined reducer in the Redux store.js.