Implemented Update User Profile facility.
In backend in userRouter.js, implemented a handler for PUT requests to
the endpoint '/api/users/profile'. This calls isAuth() middleware to
make sure only authenticated users can update profiles using this
endpoint. This handler then looks up the user with the ID sent in the
request in the database, and if such a user is found, then it does the
following: it updates the info of the user object which we got from
the database with the name, email, and password sent in the reqest. The
password is encrypted before being stored. It then saves this user in
the database. Finally it sends this user object in its response to the
frontend.

We created a new React component, <PrivateRoute> which behaves
similarly to the <Route> component, except that it checks whether the
user is authenticated by signing in, and if so then it renders just
like the <Route> component with the same props. But if no user is
logged in, then it redirects to the endpoint '/signin'.

Changed the <Route> for '/profile' in the frontend to be a
<PrivateRoute>. This fixes an issue in which when viewing the user
profile screen, if the user logs out, then an error occurs. We avoid
the error by redirecting to the /signin screen if the user has logged
out.

Implemented the updateUserProfile action, along with the concommitant
constants and the userUpdateProfileReducer reducer function in
userReducers.js. We added the new userUpdateProfile reducer to the
combined reducers in the Redux store.js.

Modified the Profile Screen so that when the Update button is clicked,
the submit handler dispatches the updateUserProfile action passing in
the info in the form fields. Also added state to keep track of the
name, email and password values.