Began implementing Admin view of the app, including Admin view of all products.
In the backend, added a new middleware function to utils.js, isAdmin(), that
checks if the user is an admin user and if so invokes the next middleware
function in the middleware stack, otherwise sending an error to the frontend.
This is closely modeled on the isAuth() middleware function.

Created a new <AdminRoute> React component to create a guarded route in the
frontend, similar to the <PrivateRoute> component. <AdminRoute> checks if the
logged in user is an admin user and if so renders the path, otherwise
redirecting to the '/signin' endpoint.

In App.js, added a new Admin dropdown menu to the right end of the header which
only displays if the user is an authenticated admin user. One of the entries in
this menu links to /productlist, and we added an AdminRoute for this path
handled by the new ProductListScreen component.

Implemented a new ProductListScreen which is only accessible to admin users, to
view all products in the database. This screen uses the same listProducts action
and reducer that we implemented for the Home Screen.