Implemented part of create product capability.
In backend, implemented a handler for POST requests to the
'/api/products' endpoint, which will be for creating a new product.
For now, this handler just creater a new product with sample (i.e.
dummy) data, saves it in the database and sends it to the frontend.

In App.js, modified the <Route path="/product/:id"> tag to add 'exact'
attribute. This will prevent this route from responding to the path
'/product/:id/edit', which is for editing a product and will be added
later.

Implemented a new createProduct action, along with concommitant
constants and the productCreateReducer reducer.

Added productCreateReducer to the combined reducer in the Redux store.

In the ProductListScreen, added a Create Product button to the header.
Also added some code to get state from the Redux store related to the
product creation process. In the useEffect hook, if product was
created successfully, dispatch PRODUCT_CREATE_RESET, and redirect to
the product edit URL for the created product.