Built the product edit screen.
Implemented most of ProductEditScreen component, for editing product
details. This screen dispathes the detailsProduct action to look up
the details of the product with the given ID from the backend. It
fills all the fields of the form on the screen with the product's
info.

The form on this page has an Update button, which when clicked calls
the submitHandler function. At this point, the submitHandler function
is just a stub; we will expand it later to update the product info in
the database.

In App.js, added a new AdminRoute for the path /product/:id/edit,
which is handled by the ProductEditScreen. NOTE: in the Amazona
course, Basir made this a regular route, not an AdminRoute, as I am
doing here. I think it makes more sense for this to be an AdminRoute,
because it should not be accessible to non-Admin users.