Completed implementing updating a product's info.
In the backend, implemented a handler for PUT requests to the endpoint
/api/products/:id. This only works if isAuth and isAdmin are true, and
it stores the updated product fields in the database. If successful,
it returns the updated product to the frontend.

In the frontend, we implemented a new updateProduct action, along with
the concommitant constants and the productUpdateReducer reducer. The
updateProduct action sends the AJAX PUT request to the backend and
dispatches the related actions to the Redux store. We added the
productUpdateReducer to the combined reducer in the Redux store.

We modified the ProductEditScreen so that the submitHandler function
for the update form dispatches the updateProduct action with the
information in the form fields. We also extracted new local variables
loadingUpdate, successUpdate and ErrorUpdate from the Redux store and
modified the useEffect() function so that if successUpdate is true, we
clear the success field and dispatch detailsProduct so the product
object fields get the updated values.