Began implementing uploading product image files to the server.
Created /uploads/ directory in the project root folder, to hold
uploaded product images. Added an empty file named file.txt to the
/uploads directory so that the uploads directory will be stored in the
git repository. Added an entry to .gitignore, /uploads/*.jpg, so that
the image files within this directory will NOT be stored in the git
repository.

Added a new frontend/public/uploads/ directory as an experiment. I
wanted to see if product image files in this directory would appear in
the product listing screen, and THEY DO. So this could be an alternate
way of storing the product image files, rather than in the /uploads
folder in the project root. For now, we are using the method Basir
uses and storing images in /uploads in the project root, using some
extra code in the server.js file for serving static files from that
folder. So I renamed frontend/public/uploads to
frontend/public/uploads.BAK, to move it out of the way.

In the root project, installed via npm the package multer, which is
used for uploading files to the server.