Finished implementing uploading product image files to the server.
In backend, implemented a new router, uploadRouter.js, that handles
POST requests to the '/api/uploads/' endpoint. This uses isAuth
middleware to ensure only authenticated users can use this function.
It uses multer middleware to upload a single file named 'image' to the
server, and using multer's diskStorage function, to store it on the
server's disk in the /uploads folder, with a unique name consisting of
the current timestamp.

In backend server.js, mounted the new uploadRouter on the path
/api/uploads. We also added a new call to app.use() that makes it so
that files with paths starting with /uploads/... will serve static
files from the /uploads directory. This way product image files will
be served from the /uploads folder in the project root, instead of
from the frontend/public/images folder, as they were before.

Modified the ProductEditScreen to add a Choose File input to the form,
along with an uploadFileHandler function to allow the user to choose a
file on their computer, and upload it to the server with an AJAX
request (via axios) to the /api/uploads endpoint. If the upload is
successful, the image file is given a unique name of the current
timestamp and put in the /uploads folder, and the server returns to
the frontend the path of this new file on the server, which becomes
the product's image attribute.