Implemented delete product functionality on Admin product list screen.
In the backend, in productRouter.js, implemented a handler for DELETE
requests to the endpoint /api/products/:id. This handler calls the
isAuth and isAdmin middleware functions to ensure that only
authenticated, admin users can use this endpoint. It looks up the
product by its id (in the request URL) in the database, and if found
it calls product.remove() to delete the product from the database and
then returns a message to the frontend indicating success and with a
copy of the deleted product. If the product is not found in the
database, it returns a 404 response code to the frontend with the
message, 'Product not found.'

Implemented the deleteProduct action, along with the concommitant
constants and the productDeleteReducer reducer.
We added the productDeleteReducer to the combined reducer in the Redux
store, store.js.

In the ProductListScreen, we finished implementing the deleteHandler()
function which is called when the Delete button is clicked next to a
product in the list. This function opens a confirmation modal dialog
and if the user confirms the deletion request, it dispatches the
deleteProduct action.

In this component also, we added code to get the product delete state
from Redux and set the local loadingDelete, successDelete and
errorDelete variables.

In the ProductListScreen we modified the useEffect() function so that,
if deletion is successful, we dispatch a PRODUCT_DELETE_RESET action.
This removes the successDelete condition to prevent an infinite loop.