Implemented list orders Admin screen, for Admin to view all orders.
In the backend, orderRouter.js, we implemented a handler for GET
requests to the /api/orders endpoint. This uses isAuth and isAdmin so
that only authenticated Admin users can use this function. It calls
Order.find({}).populate('user', 'name') to get all orders from the
database. The populate method is like a SQL JOIN. The user field in
the Order schema is a mongoose ObjectId linked to the User collection
in the database. We populate this with data from the User collection,
but in this case, only the 'name' field is pulled from User.

In App.js, added an <AdminRoute> for the path /orderlist which is
handled by the OrderListScreen component.

We implemented the listOrders action, along with the concommitant
constants and the orderListReducer reducer.The listOrders action
dispatches ORDER_LIST_REQUEST and then uses an AJAX request to
/api/orders to get all the orders and return them to the frontend, or
an error.

In the Redux store, added orderListReducer to the combined reducer.

Implemented the OrderListScreen component. This is very similar to the
screen which displays a user's orders. In the useEffect() hook, it
dispatches the listOrders action to get all the orders from the
backend, and uses the orderList reducer to get the full list of all
orders, rather than the orders made by a particular user.

All the orders are displayed in a table, which here has an additional
column for the user, so the Admin viewing this can see who made each
order. There are Depails and Delete buttons on each row. The Details
button uses props.history.push() to send the user to the order details
screen for that particular order. The Delete button doesn't function
yet. That will be implemented later.