Implemented deleting orders in the Admin list all orders screen.
In the backend, we implemented a handler for DELETE requests to the
route /api/orders/:id. This route handler uses the isAuth and isAdmin
middleware functions to ensure that only authenticated admin users can
access this function.

We implemented the deleteOrder action, along with the concommitant
constants, including RESET. This action, as usual, dispatches an
ORDER_DELETE_REQUEST then uses axios.delete() function to send the
AJAX request to the backend to delete the order with the given id. If
successful, it then dispatches ORDER_DELETE_SUCCESS, or if not
dispatches ORDER_DELETE_FAIL.

We implemented the orderDeleteReducer reducer, which operates just
like the other success:true reducers. We then added orderDeleteReducer
to the combined reducer in the Redux store.

In the OrderListScreen, we added conditional rendering to the return
value, which if loadingDelete is true renders a <LoadingBox> and if
errorDelete is true renders a <MessageBox> with the error. We also set
loadingDelete, errorDelete and successDelete via pattern matching from
orderDelete, which we select from the state in the Redux store.

In the deleteHandler function, we get confirmation from the user that
they really want to delete the order, and if so then we dispatch the
deleteOrder action with the order id.

In useEffect, we check if successDelete is true, and if so we dispatch
ORDER_DELETE_RESET, to remove the success: true field from the object.