Implemented registering an order as delivered.
In the backend, implemented a handler for the route
/api/orders/:id/deliver. This function uses isAuth and isAdmin to
ensure only authenticated admin users are able to access this route.
It looks up the order by id in the database, and if found it sets the
isDelivered field of the order document to true and the deliveredAd
field to the current date and time. It returns the updated order to
the frontend with the 'Order updated' message.

Implemented the deliverOrder action, along with the concommitand
constants. This action works very similarly to the order pay action,
dispatching ORDER_DELIVER_REQUEST with the order id, and then using
the axios.put() function to send the AJAX request to
/api/orders/${orderId}/deliver. If this is successful it then
dispatches ORDER_DELIVER_SUCCESS, and if there's a problem it
dispatches ORDER_DELIVER_FAIL.

We implemented the orderDeliverReducer reducer. This works exactly
like the orderPayReducer, including ORDER_DELIVER_RESET. This reducer
was added to the combined reducer in the Redux store.

We modified the OrderScreen to add a conditional rendering section
that renders a Deliver Order button if is the user is an admin and the
order is paid and the order has not yet been delivered. We implemented
the deliverHandler() function that is called when the Deliver Order
button is clicked. This dispatches the deliverOrder action.

Also in the OrderScreen, we set loadingDeliver, errorDeliver and
successDeliver via pattern matching from orderDeliver, which we select
from the Redux store state.

In useEffect, we added successDeliver to the if condition that if true
triggers the dispatching of detailsOrder again. We also made this if
block dispatch ORDER_DELIVER_RESET, to remove the success:true field
from the object. and added successDeliver to the list of dependencies
that cause useEffect to run again.