Made modifications for publishing the app to Heroku.
In backend/server.js, we added app.use(express.static()) to allow
Heroku to serve the React frontend files from the build folder in the
frontend. Also, added app.get('*', ) so GET requests to all URLs
return the index.html file in the frontend, and we removed the old
app.get('/'...) which returned the message, "Express Server at your
service."

In the package.json file, we added a new key "engines" with value
"node": "17.9.0" and "npm": "8.5.5", which specifies to Heroku the
versions of Node.js and npm that this app should use.

We also added a build script to the package.json file that will be
used by Heroku.